Update local.js
Had an issue where if an uploaded file was deleted by some other process and I had a record of it elsewhere in my mongo db and wanted to delete it, it's absence at the location caused the server to crash.  So adding the callbacks and error passing solves the issue for me.  Hope you can pull this request.